requirements.txt
coverage==6.3.2
pytest==7.0.1
termcolor==1.1.0 
pymongo==4.0.2
alive-progress==2.3.1
emoji==1.6.3
console-menu==0.7.0
python-dotenv==0.19.2 